requirements.txt
# Deep Research - Requirements
gradio>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
openai-agents>=0.1.0
sendgrid>=6.0.0
aiohttp>=3.9.0
